pyproject.toml
[tool.poetry]
name = "sensorial"
version = "0.5.0"
description = "API for get sensory information"
authors = ["Rodolfo De Nadai <rdenadai@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.11.0"
httpx = "^0.23.0"
pyephem = "^9.99"
fastapi = {version = "0.87.0", extras = ["all"]}
gunicorn = "^20.1.0"
boto3 = "^1.26.11"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
pylint = "^2.13.5"
mypy = "^0.942"
mccabe = "0.6.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
perflint = "^0.7.1"
debugpy = "^1.6.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = ["app"]